Merge branch 'fix-ctrl-c-win' of https://github.com/simonratner/incubator-mynewt-core into develop

+++ hw/scripts/openocd.sh
@@ -65,37 +65,51 @@ openocd_debug () {
     echo "$EXTRA_JTAG_CMD" >> $OCD_CMD_FILE
 
     if [ -z "$NO_GDB" ]; then
-	if [ -z $FILE_NAME ]; then
-	    echo "Missing filename"
-	    exit 1
-	fi
-	if [ ! -f "$FILE_NAME" ]; then
-	    echo "Cannot find file" $FILE_NAME
-	    exit 1
-	fi
+        if [ -z $FILE_NAME ]; then
+            echo "Missing filename"
+            exit 1
+        fi
+        if [ ! -f "$FILE_NAME" ]; then
+            echo "Cannot find file" $FILE_NAME
+            exit 1
+        fi
 
-	#
-	# Block Ctrl-C from getting passed to openocd.
-	#
-	set -m
-	openocd $CFG -f $OCD_CMD_FILE -c init -c halt &
-	set +m
+        if [ -x "$COMSPEC" ]; then
+            #
+            # Launch openocd in a separate command interpreter, to make sure
+            # it doesn't get killed by Ctrl-C signal from bash.
+            #
+            $COMSPEC /C "start $COMSPEC /C openocd $CFG -f $OCD_CMD_FILE -c init -c halt"
+        else
+            #
+            # Block Ctrl-C from getting passed to openocd.
+            #
+            set -m
+            openocd $CFG -f $OCD_CMD_FILE -c init -c halt &
+            set +m
+        fi
 
-    	GDB_CMD_FILE=.gdb_cmds
+        GDB_CMD_FILE=.gdb_cmds
 
-    	echo "target remote localhost:3333" > $GDB_CMD_FILE
-	if [ ! -z "$RESET" ]; then
-    	    echo "mon reset halt" >> $GDB_CMD_FILE
-	fi
-	arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME
-	rm $GDB_CMD_FILE
+        echo "target remote localhost:3333" > $GDB_CMD_FILE
+        if [ ! -z "$RESET" ]; then
+            echo "mon reset halt" >> $GDB_CMD_FILE
+        fi
+        if [ -x "$COMSPEC" ]; then
+            $COMSPEC /C "start $COMSPEC /C arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME"
+        else
+            arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME
+            rm $GDB_CMD_FILE
+        fi
     else
-	# No GDB, wait for openocd to exit
-	openocd $CFG -f $OCD_CMD_FILE -c init -c halt
-	return $?
+        # No GDB, wait for openocd to exit
+        openocd $CFG -f $OCD_CMD_FILE -c init -c halt
+        return $?
     fi
 
-    rm $OCD_CMD_FILE
+    if [ ! -x "$COMSPEC" ]; then
+        rm $OCD_CMD_FILE
+    fi
     return 0
 }
